requirements.txt
flask==2.3.2
requests==2.31.0
spacy==3.6.1
scikit-learn==1.3.0
python-dotenv==1.0.0
pydantic==1.10.9